Installation Integration Tests
This PR introduces the first of integration tests for the installation of StackStorm. This is needed for a plethora of reasons at this point:

We have many install methods (Chef/Puppet/st2_deploy.sh), and they are already starting to drift
Asserting success on installation is a bit on a hope and a prayer today (Happy Path). This actually gets us some confidence in what's making it to end-machine
Needed to start a build pipeline. In order to start packaging up Vagrant into a package that we can distribute a bit more shrink-wrapped, need to ensure that we're wrapping the right stuff
Confidence migrating to st2workroom for 0.10 release.
Docker Build. In order to properly build out docker containers, need to have some baseline.
This PR models st2_deploy.sh, with a passing grade. This does not pass Puppet installation (some defaults, and auth not being enabled), but this is good... there is now a Red/Green Refactor cycle.