Fix mistral inquiry test
Fix asserts in the mistral inquiry workflows. Change the options for ths `st2 execution get` commands from `-j` to `-dj` which will return a valid JSON. Instead of using grep and hoping the result to be somewhere, use python to load the JSON string and look for the result in exact location.